[feat]: 경로 생성 기능 추가

+++ backend/src/main/java/io/ssafy/p/j9b304/backend/domain/walk/dto/response/WalkInitialInfoResponseDto.java
@@ -23,7 +23,7 @@ public class WalkInitialInfoResponseDto {
     Integer estimatedTime;
 
     @Schema(description = "예상 이동 거리(km)")
-    Float estimatedDistance;
+    Double estimatedDistance;
 
     @Schema(description = "테마명")
     String themeName;
